% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{chng_vals}
\alias{chng_vals}
\title{Calculate Change variables}
\usage{
chng_vals(
  df,
  colname = c("covid_hospital_admissions_per_100k", "covid_inpatient_bed_utilization",
    "covid_cases_per_100k"),
  chngstr = c("chng_hospadmit", "chng_inpatient", "chng_covid"),
  .countycol = county,
  .dateformat = date_updated
)
}
\arguments{
\item{df}{CDC community level dataframe processed by format_date}

\item{colname}{Name or vector of names of columns to calculate change in previous week for as strings. Default c("covid_hospital_admissions_per_100k", "covid_inpatient_bed_utilization", "covid_cases_per_100k")}

\item{chngstr}{Name or vector of names of new column as string. Default c("chng_hospadmit", "chng_inpatient", "chng_covid")}

\item{.countycol}{Data-masked name of county column; default count}

\item{.dateformat}{Data-masked name of date column; default date_updated}
}
\value{
Dataframe with change in
}
\description{
Calculates change in a given variable or variables from previous week to next week, as well as percentage changes.
}
\examples{
{
df <- cdccovidplotting::read_county_commlevels("Maryland")
df <- cdccovidplotting::format_date(df)

dfone <- cdccovidplotting::chng_vals(df, colname = 'covid_cases_per_100k', "cases")

dftwo <- cdccovidplotting::chng_vals(df, colname = c("covid_cases_per_100k", "covid_hospital_admissions_per_100k"), c("cases", "hospadmit"))

dfthree <- cdccovidplotting::chng_vals(df)

}
}
