% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{format_date}
\alias{format_date}
\title{Format Date Columns}
\usage{
format_date(df, datecol = date_updated)
}
\arguments{
\item{df}{CDC community-level dataset}

\item{datecol}{Name of datecolumn formatting; default date_updated}
}
\value{
Dataframe with formatted date_updated and date_format columns
}
\description{
Formats date columns in CDC community level dataset, outoutting one column as date_format (in Month day, Year format) and one as date_updated (in Month-day format)
}
\examples{
{
df <- cdccovidplotting::read_county_commlevels("Maryland")

format_date(df)

}
}
