% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lines_approp}
\alias{lines_approp}
\title{Add CDC Community Level Threshold Lines}
\usage{
lines_approp(
  baseplot,
  varstring,
  countydf,
  .covidvar = "covid_cases_per_100k",
  .countycol = "county"
)
}
\arguments{
\item{baseplot}{plotly plot of CDC community level data}

\item{varstring}{Name of variable reflected in plot as string; determines at which levels to draw thresholds for}

\item{countydf}{CDC community-level dataframe filtered to one county}

\item{.covidvar}{String name of variable associated with covid-cases per 100,000 residents; default covid_cases_per_100k}

\item{.countycol}{String name of variable associated with filtered county; default county}
}
\value{
plotly plot with threshold dotted lines added to it for supplied variable
}
\description{
Adds dotted lines for thresholds at which CDC community level increases
}
\examples{
{

df <- cdccovidplotting::read_process_commlevels("Maryland")

df_county <- dplyr::filter(df, county == "Allegany County")

plot <- cdccovidplotting::plot_commlevels(vareval = ~ covid_hospital_admissions_per_100k, subtitle = "Hospital admits", df_county, basechng = "chng_hospadmit")
lines_approp(plot, "covid_hospital_admissions_per_100k", df_county)

}
}
