% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_commlevels}
\alias{plot_commlevels}
\title{Plot Community-level Components}
\usage{
plot_commlevels(countydf, vareval, subtitle, basechng)
}
\arguments{
\item{countydf}{CDC community-level dataframe filtered to one county}

\item{vareval}{Expression of variable plotting in y-axis. E.g., ~ covid_cases_per_100k.}

\item{subtitle}{Title of plot}

\item{basechng}{String-name of variable measuring change from previous week associated with vareval. E.g., "chng_covid" would be associated with the evaluated variable ~ covid_cases_per_100k. Assumes that the percentage-change variable shares the same base-name}
}
\value{

}
\description{
Creates plotly plot of metrics that compose CDC community-level
}
\examples{
{

df <- cdccovidplotting::read_process_commlevels("Maryland")

df_county <- dplyr::filter(df, county == "Allegany County")

plot_commlevels(vareval = ~ covid_hospital_admissions_per_100k, subtitle = "Hospital admits", df_county, basechng = "chng_hospadmit")

}
}
