% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_commlevels_lines}
\alias{plot_commlevels_lines}
\title{Plot Community-level data and draw threshold lines}
\usage{
plot_commlevels_lines(
  countydf,
  vareval,
  subtitle,
  basechng,
  countyfilt = NULL,
  .covidvar = covid_cases_per_100k,
  .countycol = county
)
}
\arguments{
\item{countydf}{CDC community-level dataframe filtered to one county. Alternatively, can provide unfiltered dataframe with countyfilt set to county want to filter data to}

\item{vareval}{Expression of variable plotting in y-axis. E.g., ~ covid_cases_per_100k.}

\item{subtitle}{Title of plot}

\item{basechng}{String-name of variable measuring change from previous week associated with vareval. E.g., "chng_covid" would be associated with the evaluated variable ~ covid_cases_per_100k. Assumes that the percentage-change variable shares the same base-name}

\item{countyfilt}{String name of county to filter countydf to; default NULL, which assumes countydf is already filtered to one county. Performs grepl match with countyfilt so partial matches acceptable}

\item{.covidvar}{Data-masked name of variable associated with covid-cases per 100,000 residents; default covid_cases_per_100k}

\item{.countycol}{Data-masked name of variable associated with filtered county; default county}
}
\value{
Plot of CDC community-level COVID-19 data for one county with risk-threshold lines drawn
}
\description{
Generates plot of community-level COVID-19 data for given county and draws threshold lines at which point risk threshold increases. Wrapper for cdccovidplotting::plot_commlevels and cdccovidplotting::lines_approp functions.
}
\examples{
{

df <- cdccovidplotting::read_process_commlevels("Maryland")

df_county <- dplyr::filter( df, county == "Allegany County")

cdccovidplotting::plot_commlevels_lines(vareval = ~ covid_hospital_admissions_per_100k, subtitle = "Hospital admits", countydf = df_county, basechng = "chng_hospadmit")

cdccovidplotting::plot_commlevels_lines(vareval = ~ covid_hospital_admissions_per_100k, subtitle = "Hospital admits", countydf = df_county, basechng = "chng_hospadmit", countyfilt = "Allegany")



}
}
