% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_county_commlevels}
\alias{read_county_commlevels}
\title{Read-in CDC Community-level Data}
\usage{
read_county_commlevels(statename)
}
\arguments{
\item{statename}{Name of state to read county data for}
}
\value{
Dataset on CDC community-levels, filtered to supplied state
}
\description{
Read in county-level CDC community-level data from CDC API, filtering to provided statename
}
\examples{
{
read_county_commlevels("Maryland")

read_county_commlevels("Massachusetts")

}
}
