% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_process_commlevels}
\alias{read_process_commlevels}
\title{Read and process CDC Community-level Data}
\usage{
read_process_commlevels(statename)
}
\arguments{
\item{statename}{Name of state to process county-level information on}
}
\value{
Dataframe with formatted date columns and change in values from prior week to current week
}
\description{
Reads and processes CDC community-level data for counties, formatting date variables and calculating changes in variables from each week
}
\examples{
{
read_process_commlevels("Maryland")

read_process_commlevels("Alabama")

}
}
