% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{subplots_keyvars}
\alias{subplots_keyvars}
\title{Generate subplots of Community-level components}
\usage{
subplots_keyvars(
  countydf,
  varevals = c(~covid_cases_per_100k, ~covid_hospital_admissions_per_100k,
    ~covid_hospital_admissions_per_100k),
  subtitles = c("COVID cases\\nper 100k",
    "Percent of beds\\nfilled by\\nCOVID patients", "COVID hospital\\nadmits per\\n100k"),
  basechngs = c("chng_covid", "chng_inpatient", "chng_hospadmit"),
  countyfilt = NULL,
  subrows = 2,
  .covidvar = covid_cases_per_100k,
  .countycol = county
)
}
\arguments{
\item{countydf}{CDC community-level dataframe filtered to one county. Alternatively, can provide unfiltered dataframe with countyfilt set to county want to filter data to}

\item{varevals}{Expression of variables plotting in y-axis. E.g., ~ covid_cases_per_100k. Default c(~ covid_cases_per_100k, ~ covid_hospital_admissions_per_100k, ~ covid_hospital_admissions_per_100k)}

\item{subtitles}{Titles of subplots. Default c("COVID cases\ per 100k", "Percent of beds filled by COVID patients", "COVID hospital admits per 100k")}

\item{basechngs}{String-names of variables measuring change from previous week associated with vareval. E.g., "chng_covid" would be associated with the evaluated variable ~ covid_cases_per_100k. Assumes that the percentage-change variable shares the same base-name. Default c("chng_covid", "chng_inpatient", "chng_hospadmit")}

\item{countyfilt}{String name of county to filter countydf to; default NULL, which assumes countydf is already filtered to one county. Performs grepl match with countyfilt so partial matches acceptable}

\item{subrows}{Number of rows in subplot; default 2}

\item{.covidvar}{Data-masked name of variable associated with covid-cases per 100,000 residents; default covid_cases_per_100k}

\item{.countycol}{Data-masked name of variable associated with filtered county; default county}
}
\value{
Subplots of COVID-19 variables
}
\description{
Generates subplot of plotly plots of component-metrics that make up CDC community level for one county.
}
